Rename MazePainter3D to MazePainterMap and move to Painters.Map namespace
